moving settings around v2
