+++ src/powerquery-parser/common/index.ts
@@ -13,6 +13,7 @@ export * as Traverse from "./traversal";
 export * as TypeScriptUtils from "./typeScriptTypeUtils";
 
 export * from "./cancellationToken";
+export * from "./commonSettings";
 export * from "./immutableSet";
 export * from "./orderedMap";
 export * from "./partialResult";
