+++ src/powerquery-parser/language/index.ts
@@ -9,5 +9,5 @@ export { Comment, Token };
 export * from "./ast";
 export * from "./constant";
 export * from "./keyword";
-export * from "./settings";
+export * from "../lexer/lexSettings";
 export * from "./type";
