+++ src/powerquery-parser/lexer/lexSettings.ts
@@ -1,0 +1,6 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+import { CommonSettings } from "../common";
+
+export type LexSettings = CommonSettings;
